+++ source/Burntime.Remaster/Helper.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-using Burntime.Framework;
+﻿using Burntime.Platform.Graphics;
 using Burntime.Platform.Resource;
-using Burntime.Platform.Graphics;
 
 namespace Burntime.Remaster
 {
@@ -12,22 +7,63 @@ namespace Burntime.Remaster
     {
         static public DataID<ISprite> GetCharacterBody(int body, int colorSet)
         {
-            int sprite = (body + 1) * 16 + 48 * colorSet;
-            return BurntimeClassic.Instance.ResourceManager.GetData("burngfxani@syssze.raw?" + sprite + "-" + (sprite + 15));
+            int sprite = GetBodyId(body, colorSet);
+            if (sprite < 0)
+                sprite = 0;
+            return BurntimeClassic.Instance.ResourceManager.GetData("burngfxani@syssze.raw?" + sprite + "-" + (sprite + 15), ResourceLoadType.LinkOnly);
         }
 
-        static public int GetSetBodyId(Burntime.Data.BurnGfx.Save.CharacterType characterType)
+        static public int GetBodyId(int body, int color)
         {
-            switch (characterType)
+            if (color < 0 || color >= 3)
+                return -1;
+
+            if (body == 3)
+                return (14 + color) * 16;
+            else if (body >= 0 && body < 3)
+                return (body + 1) * 16 + 48 * color;
+
+            return -1;
+        }
+
+        static public int GetSetBodyId(Data.BurnGfx.Save.CharacterType characterType)
+        {
+            return characterType switch
             {
-                case Burntime.Data.BurnGfx.Save.CharacterType.Mercenary:
-                    return 1;
-                case Burntime.Data.BurnGfx.Save.CharacterType.Technician:
-                case Burntime.Data.BurnGfx.Save.CharacterType.Doctor:
-                    return 0;
-                default:
-                    return 2;
-            }
+                Data.BurnGfx.Save.CharacterType.Mercenary => 1,
+                Data.BurnGfx.Save.CharacterType.Doctor => 0,
+                Data.BurnGfx.Save.CharacterType.Technician => 2,
+                Data.BurnGfx.Save.CharacterType.Boss => 3,
+                _ => -1,
+            };
+        }
+
+        static public int GetSetBodyId(CharClass charClass)
+        {
+            return charClass switch
+            {
+                CharClass.Mercenary => 1,
+                CharClass.Doctor => 0,
+                CharClass.Technician => 2,
+                CharClass.Boss => 3,
+                _ => -1,
+            };
+        }
+
+        static public int GetColorFromSpriteId(int spriteId)
+        {
+            spriteId /= 16;
+
+            if (spriteId > 16)
+                return -1;
+
+            if (spriteId >= 14)
+                return spriteId - 14;
+
+            if (spriteId >= 1 || spriteId <= 9)
+                return (spriteId - 1) / 3;
+
+            return -1;
         }
     }
 }